% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHitScore.R
\name{calcHitScore}
\alias{calcHitScore}
\title{Calculate hit score that summarizes the biallelic state of a variant pair or gene}
\usage{
calcHitScore(df, mode, min.hit.score.filter = 0,
  ignore.additional.evidence = F)
}
\arguments{
\item{df}{A dataframe that is the output of mkBialleleMutProfile()}

\item{mode}{Can be: 'cnv_germ', 'cnv_som', 'germ_som'. Hit score calculation is different for
each biallele combination.}

\item{min.hit.score.filter}{Scores less than this will be set to 0.}

\item{ignore.additional.evidence}{If TRUE, will only consider the base "5+5" score}
}
\value{
The hit_score as a float
}
\description{
Calculate hit score that summarizes the biallelic state of a variant pair or gene
}
