% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHitScores.R
\name{calcHitScores}
\alias{calcHitScores}
\title{Sums the pathogenicity score of allele 1 and 2 (hit score)}
\usage{
calcHitScores(gene.diplotypes, diplotype.origin.rank = c("cnv_cnv",
  "cnv_som", "cnv_germ", "germ_som"))
}
\arguments{
\item{gene.diplotypes}{Table containing the mutation info of alleles 1 and 2}

\item{diplotype.origin.rank}{A character vector of diplotype origins in descending order
of pathogenic impact}
}
\value{
A data frame containing the hit scores and boosted hit scores (scores are boosted by
order of diplotype.origin.rank to maintain priority when determining the max diplotype effect)
}
\description{
Sums the pathogenicity score of allele 1 and 2 (hit score)
}
