% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessDiplotypes.R
\name{detHitType}
\alias{detHitType}
\title{Determine biallelic hit type}
\usage{
detHitType(diplotypes, min.a1.score = 1, min.a2.score = 1)
}
\arguments{
\item{diplotypes}{Diplotypes dataframe}

\item{min.a1.score}{Min score to assess hit type}

\item{min.a2.score}{See above}
}
\value{
A character vector of hit types
}
\description{
Determine biallelic hit type
}
