% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postProcessDiplotypes.R
\name{detReadingFrame}
\alias{detReadingFrame}
\title{Determines the reading frame in case of one or more frameshifts}
\usage{
detReadingFrame(gene.diplotypes, mut.profile.germ, mut.profile.som)
}
\arguments{
\item{gene.diplotypes}{Table containing the mutation info of alleles 1 and 2}

\item{mut.profile.germ}{Germ mut profile table}

\item{mut.profile.som}{Som mut profile table}
}
\value{
A dataframe containing info of: the number of detected indels, the lengths of each
frameshift, the reading frame (sum of the frameshift lengths), and a logical indicating whether
the gene is out of frame
}
\description{
Determines the reading frame in case of one or more frameshifts
}
