% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneDiplotypes.R
\name{getGeneDiplotypes}
\alias{getGeneDiplotypes}
\title{Per gene, get row(s) where score is max}
\usage{
getGeneDiplotypes(df, mode, simplify.snpeff.eff = T)
}
\arguments{
\item{df}{A dataframe containing ensembl_gene_id}

\item{mode}{Can be: 'cnv_germ', 'cnv_som', 'germ_som'.}

\item{simplify.snpeff.eff}{Simplify snpeff_eff (e.g. stop_gained, stop_loss; becomes nonsense)}
}
\value{
A subsetting dataframe
}
\description{
Per gene, get row(s) where score is max
}
