% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneMaxEff.R
\name{getGeneMaxEff}
\alias{getGeneMaxEff}
\title{Return row(s) with the maximum effect score per gene}
\usage{
getGeneMaxEff(df, colname = "max_score", greedy.max.eff = T,
  show.n.max = F)
}
\arguments{
\item{df}{A mut_profile dataframe containing a max_score column}

\item{colname}{An colname to determine the max effect. Default='max_score'}

\item{greedy.max.eff}{Return only the one instance with maximum score per gene?}

\item{show.n.max}{Add a column indicating the number of rows with max score per gene?}
}
\value{
A subsetted dataframe
}
\description{
Return row(s) with the maximum effect score per gene
}
