% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkBialleleMutProfile.R
\name{mkBialleleMutProfile}
\alias{mkBialleleMutProfile}
\title{Performs pairwise joining of the cnv, germ, and som mut_profile tables}
\usage{
mkBialleleMutProfile(mut_profile, verbose = T)
}
\arguments{
\item{mut_profile}{mut_profile object. A list of of cnv, germ and som tables.}

\item{verbose}{Show messages?}
}
\value{
A list containing cnv_germ, cnv_som, and germ_som tables
}
\description{
Rows are joined by ENSEMBL gene id. All rows are kept (natural join).
}
