% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkDatabases.R
\name{mkClinvarTxt}
\alias{mkClinvarTxt}
\title{Download ClinVar vcf and convert to txt file}
\usage{
mkClinvarTxt(
  out.file,
  tmp.dir = "~/",
  url = "ftp://ftp.ncbi.nlm.nih.gov/pub/clinvar/vcf_GRCh37/clinvar_20200224.vcf.gz",
  java.path = JAVA_PATH,
  snpsift.path = SNPSIFT_PATH,
  verbose = T
)
}
\arguments{
\item{out.file}{Path to output file}

\item{tmp.dir}{Temporary processing directory. Defaults to ~/}

\item{java.path}{Path to java binary (defaults to the installed JRE location)}

\item{snpsift.path}{Path to SnpSift jar (defaults to the one included in this package)}

\item{verbose}{Show progress messages?}
}
\description{
Download ClinVar vcf and convert to txt file
}
