% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkGeneDiplotypes.R
\name{mkGeneDiplotypesGermSom}
\alias{mkGeneDiplotypesGermSom}
\title{Merge germ and som mut profiles into a table the describes the events of allele 1 and 2}
\usage{
mkGeneDiplotypesGermSom(mut.profile.germ, mut.profile.som, verbose = T)
}
\arguments{
\item{verbose}{Print progress messages?}

\item{mut.profile.cnv}{CNV mut profile table}

\item{mut.profile.mut}{germ or som mut profile table}

\item{mut.origin}{Indicate whether allele 2 originates from 'germ' or 'som'. Used to create
the diplotype_origin column}
}
\value{
A table the describes the events of allele 1 and 2
}
\description{
Per gene, only the most pathogenic germ/som variant pairs are merged. This
prevents the table from growing exponentially.
}
