% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkMutProfileCnv.R
\name{mkMutProfileCnv}
\alias{mkMutProfileCnv}
\title{Create annotations to the HMF CNV table}
\usage{
mkMutProfileCnv(df.cnv,
  min.cn.diff.in.gene = CUTOFFS$min.cn.diff.in.gene,
  max.max.copy.number = CUTOFFS$max.max.copy.number,
  min.min.copy.number = CUTOFFS$min.min.copy.number,
  max.min.minor.allele.ploidy = CUTOFFS$max.min.minor.allele.ploidy,
  rm.non.selected.genes = T, genes.bed = NULL, verbose = T)
}
\arguments{
\item{df.cnv}{HMF *.gene.cnv table}

\item{min.cn.diff.in.gene}{Placeholder}

\item{max.max.copy.number}{Placeholder}

\item{min.min.copy.number}{Placeholder}

\item{max.min.minor.allele.ploidy}{Placeholder}

\item{rm.non.selected.genes}{Remove genes not specified by the user in genes.bed}

\item{genes.bed}{A bed file which contains ENSEMBL gene ids}
}
\value{
The original input dataframe with annotation columns
}
\description{
Primarily used to determine full gene loss and LOH
}
