% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkMutProfileSnvIndel.R
\name{mkMutProfileSnvIndel}
\alias{mkMutProfileSnvIndel}
\title{Create annotations to the germline/somatic tables originating from HMF germline/somatic vcfs}
\usage{
mkMutProfileSnvIndel(df.snv.indel, tumor.purity, mode,
  ignore.additional.evidence = F, rm.non.selected.genes = T,
  genes.bed = NULL, gene.identifier = "ensembl_gene_id",
  keep.only.first.eff = T, verbose = T)
}
\arguments{
\item{df.snv.indel}{A germline/somatic table originating from HMF germline/somatic vcfs}

\item{tumor.purity}{Tumor purity}

\item{mode}{Can be 'germline' or 'somatic'}

\item{rm.non.selected.genes}{Remove genes not specified by the user in genes.bed}

\item{genes.bed}{A bed file which contains ENSEMBL gene ids}

\item{gene.identifier}{Can be 'snpeff_gene' or 'ensembl_gene_id'. If 'snpeff_gene', then ENSEMBL
gene ids will be retrieved}

\item{keep.only.first.eff}{Only keep first snpeff_eff (items are separated by '&')?}

\item{verbose}{Show messages?}
}
\value{
The original input dataframe with annotation columns
}
\description{
Primarily used to determine
}
