% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preProcessHmfPlOutput.R
\name{preProcessGeneCnv}
\alias{preProcessGeneCnv}
\title{Subsets PURPLE gene cnv file for genes of interest}
\usage{
preProcessGeneCnv(
  gene.cnv.file,
  out.file,
  bed.file = BED_FILE,
  genes.enst2ensg.file = GENES_ENST2ENSG,
  verbose = T
)
}
\arguments{
\item{gene.cnv.file}{Path to PURPLE gene cnv file}

\item{out.file}{Path to output file}

\item{bed.file}{Path to bed file specifying the genome intervals to keep (defaults ot the one 
stored in this package)}

\item{genes.enst2ensg.file}{Path to lookup table to convert ENSEMBL transcript to genes ids}

\item{verbose}{Show progress messages?}
}
\value{
Nothing but writes the subsetted PURPLE gene cnv txt file
}
\description{
Subsets PURPLE gene cnv file for genes of interest
}
