% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkDatabases.R
\name{retrieveHgncGeneList}
\alias{retrieveHgncGeneList}
\title{Retrieve gene identifiers from HGNC}
\usage{
retrieveHgncGeneList(
 
    hgnc.url = "https://www.genenames.org/cgi-bin/download/custom?col=gd_hgnc_id&col=gd_app_sym&col=gd_prev_sym&col=gd_aliases&col=gd_pub_ensembl_id&status=Approved&hgnc_dbtag=on&order_by=gd_app_sym_sort&format=text&submit=submit",
  out.path = NULL,
  export.as.rdata = F,
  verbose = T,
  ...
)
}
\arguments{
\item{hgnc.url}{URL to HGNC database, where the output contains the columns: HGNC ID, Approved
symbol, Previous symbols, Synonyms, ENSEMBL gene id}

\item{out.path}{Export path}

\item{export.as.rdata}{Save as RData file?}

\item{verbose}{Show messages?}

\item{...}{Arguments that can be passed to biomaRt::useMart()}
}
\value{
A dataframe of gene identifers from HGNC
}
\description{
Retrieve gene identifiers from HGNC
}
