% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detGeneStatuses.R
\name{detGeneStatuses}
\alias{detGeneStatuses}
\title{Determine gene amplifications and biallelic losses}
\usage{
detGeneStatuses(
  out.dir,
  input.file.paths = c(germ_vcf = "", som_vcf = "", gene_cnv = "", cnv = ""),
  sample.name,
  bed.file = BED_FILE,
  java.path = JAVA_PATH,
  snpsift.path = SNPSIFT_PATH,
  chrom.arm.split.method = "hmf",
  do.snpeff.ann = F,
  verbose = T
)
}
\arguments{
\item{out.dir}{Path to output dir}

\item{input.file.paths}{A list or vector supplying the path output files from the HMF pipeline.}

\item{sample.name}{Name of the sample
This list should be in the form c(germ_vcf='', som_vcf='', gene_cnv='', cnv='')}

\item{bed.file}{Path to the bed file containing the genes of interest. This bed file should also 
contain the column ensembl_gene_id}

\item{java.path}{Path the the java binary}

\item{snpsift.path}{Path to the SnpSift jar}

\item{chrom.arm.split.method}{Can be 'hmf' or 'gap'. Refer to documentation for 
calcChromArmPloidies().}

\item{do.snpeff.ann}{Annotate SNV/indels variant type with snpeff?}

\item{verbose}{Show progress messages?}
}
\value{
Writes a diplotypes table to the output dir
}
\description{
Determine gene amplifications and biallelic losses
}
