% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detGeneStatuses.R
\name{detGeneStatuses}
\alias{detGeneStatuses}
\title{Determine gene biallelic status}
\usage{
detGeneStatuses(out.dir, cnv.path, germ.path, som.path, purity.path,
  genes.bed.path, init.path)
}
\arguments{
\item{out.dir}{Directory to write output}

\item{cnv.path}{Purple gene CNV table}

\item{germ.path}{Germline variant table}

\item{som.path}{Somatic variant table}

\item{purity.path}{Purple purity table}

\item{genes.bed.path}{Bed file of gene selection}

\item{init.path}{An R script that provides the settings for this detGeneStatuses.}
}
\value{
Write the tables biall_mut_profile_*, gene_diplotypes, and gene_diplotypes_max to out.dir
}
\description{
Determine gene biallelic status
}
