% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkMutProfileCnv.R
\name{mkMutProfileCnv}
\alias{mkMutProfileCnv}
\title{Create annotations to the HMF CNV table}
\usage{
mkMutProfileCnv(df.cnv,
  full.gene.loss.max.max.copy.number = CUTOFFS$full.gene.loss.max.max.copy.number,
  trunc.max.min.copy.number = CUTOFFS$trunc.max.min.copy.number,
  loh.max.min.minor.allele.ploidy = CUTOFFS$loh.max.min.minor.allele.ploidy,
  scoring = SCORING_CNV, rm.non.selected.genes = T, genes.bed = NULL,
  verbose = T)
}
\arguments{
\item{df.cnv}{HMF *.gene.cnv table}

\item{full.gene.loss.max.max.copy.number}{The max max_copy_number for a gene to be considered to 
be completely lost (deep deletion)}

\item{trunc.max.min.copy.number}{The max min_copy_number to for a gene to be considered to have
a truncation}

\item{loh.max.min.minor.allele.ploidy}{The maximum min_minor_allele_ploidy for a gene to be
considered to have loss of heterozygosity}

\item{rm.non.selected.genes}{Remove genes not specified by the user in genes.bed}

\item{genes.bed}{A bed file which contains ENSEMBL gene ids}
}
\value{
The original input dataframe with annotation columns
}
\description{
Primarily used to determine full gene loss and LOH
}
